@@ -19,23 +19,23 @@
   "license": "MIT",
   "dependencies": {
     "autoprefixer": "6.5.3",
-    "babel-core": "6.18.2",
-    "babel-loader": "6.2.8",
+    "babel-core": "6.20.0",
+    "babel-loader": "6.2.9",
     "babel-plugin-react-display-name": "2.0.0",
     "babel-plugin-react-transform": "2.0.2",
     "babel-plugin-transform-decorators-legacy": "1.3.4",
     "babel-plugin-undeclared-variables-check": "6.8.0",
-    "babel-polyfill": "6.16.0",
+    "babel-polyfill": "6.20.0",
     "babel-preset-es2015": "6.18.0",
     "babel-preset-es2015-loose": "8.0.0",
     "babel-preset-react": "6.16.0",
     "babel-preset-stage-1": "6.16.0",
-    "babel-runtime": "6.18.0",
+    "babel-runtime": "6.20.0",
     "body-parser": "1.15.2",
-    "bootstrap-loader": "1.3.1",
+    "bootstrap-loader": "1.3.2",
     "bootstrap-sass": "3.3.7",
     "classnames": "2.2.5",
-    "css-loader": "0.26.0",
+    "css-loader": "0.26.1",
     "dotenv": "2.0.0",
     "errorhandler": "1.5.0",
     "es6-promise": "4.0.5",
@@ -46,13 +46,13 @@
     "font-awesome-webpack": "0.0.4",
     "history": "4.4.1",
     "html-webpack-plugin": "2.24.1",
-    "imports-loader": "0.6.5",
+    "imports-loader": "0.7.0",
     "isomorphic-fetch": "2.2.1",
     "jquery": "3.1.1",
     "less": "2.7.1",
     "less-loader": "2.2.3",
     "method-override": "2.3.7",
-    "node-sass": "3.13.0",
+    "node-sass": "3.13.1",
     "postcss-import": "8.1.0",
     "postcss-loader": "0.13.0",
     "react": "15.4.1",
@@ -63,13 +63,13 @@
     "react-router-redux": "4.0.7",
     "redux": "3.6.0",
     "redux-thunk": "2.1.0",
-    "resolve-url-loader": "1.6.0",
+    "resolve-url-loader": "1.6.1",
     "sass-loader": "4.0.2",
     "style-loader": "0.13.1",
     "tcomb-form": "0.9.10",
     "url-loader": "0.5.7",
     "webpack": "2.1.0-beta.15",
-    "webpack-merge": "1.0.2",
+    "webpack-merge": "1.1.1",
     "whatwg-fetch": "2.0.1",
     "yargs": "6.5.0"
   },
@@ -79,11 +79,11 @@
     "chai-as-promised": "6.0.0",
     "clean-webpack-plugin": "0.1.14",
     "enzyme": "2.6.0",
-    "eslint": "3.11.1",
+    "eslint": "3.12.0",
     "eslint-config-airbnb": "13.0.0",
     "eslint-plugin-import": "2.2.0",
     "eslint-plugin-jsx-a11y": "2.2.3",
-    "eslint-plugin-react": "6.7.1",
+    "eslint-plugin-react": "6.8.0",
     "expect": "1.20.2",
     "ignore-styles": "5.0.01",
     "istanbul": "1.0.0-alpha.2",
