@@ -3,8 +3,8 @@ from djangoreactredux.settings.staging import *  # NOQA (ignore all errors on th
 DATABASES = {
     'default': {
         'ENGINE': 'django.db.backends.postgresql_psycopg2',
-        'NAME': 'simplepay',
-        'USER': 'postgres',
+        'NAME': 'djangoreactredux_ci',
+        'USER': 'djangoreactredux',
         'PASSWORD': '',
         'HOST': 'localhost',
         'PORT': '',
