@@ -1,5 +1,3 @@
-import os
-
 from djangoreactredux.settings.base import *  # NOQA (ignore all errors on this line)
 
 
@@ -9,8 +7,12 @@ PAGE_CACHE_SECONDS = 1
 
 DATABASES = {
     'default': {
-        'ENGINE': 'django.db.backends.sqlite3',
-        'NAME': os.path.join(BASE_DIR, 'db.sqlite'),  # NOQA (ignore all errors on this line)
+        'ENGINE': 'django.db.backends.postgresql_psycopg2',
+        'NAME': 'djangoreactredux_dev',
+        'USER': 'djangoreactredux',
+        'PASSWORD': 'password',
+        'HOST': 'postgres',
+        'PORT': 5432,
     }
 }
 
