@@ -13,8 +13,12 @@ ALLOWED_HOSTS = ['*']
 
 DATABASES = {
     'default': {
-        'ENGINE': 'django.db.backends.sqlite3',
-        'NAME': os.path.join(BASE_DIR, 'db.sqlite'),  # NOQA (ignore all errors on this line)
+        'ENGINE': 'django.db.backends.postgresql_psycopg2',
+        'NAME': 'djangoreactredux_prod',
+        'USER': 'djangoreactredux',
+        'PASSWORD': 'password',
+        'HOST': 'postgres',
+        'PORT': 5432,
     }
 }
 
