@@ -40,8 +40,9 @@ export function dataFetchProtectedData(token) {
             .catch((error) => {
                 if (error && typeof error.response !== 'undefined' && error.response.status === 401) {
                     // Invalid authentication credentials
-                    error.response.json().then((data) => {
+                    return error.response.json().then((data) => {
                         dispatch(authLoginUserFailure(401, data.non_field_errors[0]));
+                        dispatch(push('/login'));
                     });
                 } else if (error && typeof error.response !== 'undefined' && error.response.status >= 500) {
                     // Server side error
@@ -52,6 +53,7 @@ export function dataFetchProtectedData(token) {
                 }
 
                 dispatch(push('/login'));
+                return Promise.resolve(); // TODO: we need a promise here because of the tests, find a better way
             });
     };
 }
